Move resetLocked to Pool
